Modificación en el MER y base de datos. Pruebas con todos los formularios realizados y probados.

+++ PackageDelivery.Repository.Implementation/Mappers/Parameters/DocumentTypeRepositoryMapper.cs
@@ -29,7 +29,7 @@ namespace PackageDelivery.Repository.Implementation.Mappers.Parameters
         {
             return new tipoDocumento()
             {
-                id = input.Id,
+                id = (int)input.Id,
                 nombre = input.Name.Trim()
             };
         }
